// Glo/src/types/types.ts
export type OpenAIModel = 'gpt-3.5-turbo' | 'gpt-4';

export interface EssayBody {
  topic: string;
  paragraphs: 3 | 4 | 5;
  essayType:
    | ''
    | 'Argumentative'
    | 'Classic'
    | 'Persuasive'
    | 'Memoir'
    | 'Critique'
    | 'Compare/Contrast';
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface ChatBody {
  messages: { role: string; content: string }[];
  userId: string | undefined | null;
  conversationId?: string | number | null;
  userPromptId?: string | number | null;
  promptId?: string | number | null;
}

export interface SimplifierBody {
  content: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface ProductDescriptionBody {
  name: string;
  keyBenefitsFeatures: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface EmailEnhancerBody {
  topic: string;
  toneOfVoice:
    | ''
    | 'Formal'
    | 'Informal'
    | 'Humorous'
    | 'Serious'
    | 'Optimistic'
    | 'Motivating'
    | 'Respectful'
    | 'Assertive'
    | 'Conversational';
  content: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface LinkedinBody {
  topic: string;
  audience: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface CaptionBody {
  topic: string;
  toneOfVoice:
    | ''
    | 'Formal'
    | 'Informal'
    | 'Humorous'
    | 'Serious'
    | 'Optimistic'
    | 'Motivating'
    | 'Respectful'
    | 'Assertive'
    | 'Conversational';
  description: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface FaqBody {
  topic: string;
  productType: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface NameGeneratorBody {
  topic: string;
  productType: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface SeoKeywordsBody {
  name: string;
  topics: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface ReviewResponderBody {
  review: string;
  name: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface BusinessGeneratorBody {
  topic:
    | ''
    | 'Art and Entertainment'
    | 'Business Equipment and Supplies'
    | 'Clothing and Accessories'
    | 'Food and Drink'
    | 'Hardware and Automotive'
    | 'Health and Beauty'
    | 'Home and Garden'
    | 'Internet and Technology'
    | 'Pet supplies'
    | 'Sports and Recreation'
    | 'Toys and Games'
    | 'Travel & Hospitality';
  productType: '' | 'Physical' | 'Digital' | 'Service';
  budget:
    | ''
    | 'Under $500'
    | '$500-$1000'
    | '$1000-$5000'
    | '$5000-$20,000'
    | '$20,000+';
  model: OpenAIModel;
  apiKey?: string | undefined;
}
export interface ArticleBody {
  topic: string;
  title: string;
  language:
    | ''
    | 'English'
    | 'Chinese'
    | 'Spanish'
    | 'Arabic'
    | 'Hindi'
    | 'Italian'
    | 'Portuguese'
    | 'Russian'
    | 'Japanese'
    | 'Romanian'
    | 'German';
  words: 200 | 300 | 400 | 500 | 600;
  model: OpenAIModel;
  apiKey?: string | undefined;
}

export interface PlagiarismCheckerBody {
  content: string;
  model: OpenAIModel;
  apiKey?: string | undefined;
}